Avoid rotating icons more than 180 degrees
Previously, rotating the device counter-clockwise from upright
(from 0 degrees to 270 degrees) would cause the icons to spin -270
degrees to reach the correct orientation. This change compares the
previous and current device orientation to find the shortest rotation
needed to bring the icons into the correct orientation. Now, when
rotating the device from 0 degrees to 270 degrees, the icons only rotate
90 degrees as expected.