Remove and simplify image quality settings code
In the past there used to be +/- buttons for the JPEG image quality, but it was removed. For some reason the buttons still exist but are simply not visible (gone) and there is still code handling it despite it being impossible to use. This just removes it and reduces a bit of the existing code.

Signed-off-by: June <june@eridan.me>