bugfix: camConfig shouldn't be a static field of the MainActivity
When another instance of MainActivity (or its subclass) starts, it overwrites
this field, which breaks all other instances of MainActivity (or its subclass).

Make camConfig an instance field to resolve this issue.

Note that passing camConfig to MoreSettings is a bit hacky, but a proper fix
would be a lot more work for little upside.