work around R8-related regression in CameraX 1.5.0-alpha05
CameraX 1.5.0-alpha05 causes crashes when capturing photos on certain
devices. So far, the reports are all from 6th and 7th generation Pixels
which have not been updated from Android 13. We replicated the crash on
a Pixel 7a running the final Android 13 release (TQ3A.230901.001). It's
unlikely that the issue is actually specific to Pixels but rather they
may be the only devices with the relevant feature available on Android
13. Pixels are the only devices shipping monthly and quarterly Android
releases for their stock OS in practice which may be related.